Change name tmp to MainView and MainView to MainViewOld
Renaming tmp to main in files and references and the same for renaming MainView to MainViewOld
