+++ Veles Application/Views/MainViewOld.xaml
@@ -1,0 +1,40 @@
+﻿<Window x:Class="Veles_Application.MainViewOld"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:Veles_Application"
+        xmlns:viewModel="clr-namespace:Veles_Application.ViewModels"
+        mc:Ignorable="d"
+        Title="MainWindow" Height="450" Width="800">
+
+    <Window.DataContext>
+        <viewModel:MainViewModel/>
+    </Window.DataContext>
+    
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="*" />
+            <ColumnDefinition Width="*" />
+            <ColumnDefinition Width="*" />
+        </Grid.ColumnDefinitions>
+
+        <Grid.RowDefinitions>
+            <RowDefinition Height="4*" />
+            <RowDefinition Height="*" />
+            <RowDefinition Height="2*" />
+            <RowDefinition Height="*" />
+        </Grid.RowDefinitions>
+
+        <Border Grid.Row="0" Grid.Column="1" Background="GhostWhite" BorderBrush="Gainsboro" BorderThickness="1">
+            <ScrollViewer>
+                <TextBlock Text="{Binding Text}"  TextWrapping="Wrap"/>
+            </ScrollViewer>
+        </Border>
+        
+        <TextBlock Text="{Binding Username}" Grid.Row="1" Grid.Column="1" HorizontalAlignment="Left" />
+        <TextBox Text="{Binding Message, UpdateSourceTrigger=PropertyChanged}"  Grid.Row="2" Grid.Column="1"/>
+        <Button Command="{Binding SendCommand}" Content="Send" Grid.Row="3" Grid.Column="1"/>
+    </Grid>
+    
+</Window>
