--- Veles Application/Views/tmp.xaml
@@ -1,52 +1,0 @@
-﻿<Window x:Class="Veles_Application.Views.tmp"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:i="http://schemas.microsoft.com/xaml/behaviors"
-        xmlns:local="clr-namespace:Veles_Application.Views"
-        mc:Ignorable="d"
-        Title="tmp" Height="450" Width="800">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="*"/>
-            <ColumnDefinition Width="3*"/>
-            <ColumnDefinition Width="*"/>
-        </Grid.ColumnDefinitions>
-        
-
-        <Grid Grid.Column="0">
-            <Grid.RowDefinitions>
-                <RowDefinition Height="*"/>
-                <RowDefinition Height="9*"/>
-            </Grid.RowDefinitions>
-
-            <Grid Grid.Row="0">
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="*"/>
-                    <ColumnDefinition Width="*"/>
-                </Grid.ColumnDefinitions>
-
-                <Button Grid.Column="0" Content="Home" Command="{Binding ChangePanelCommand}" CommandParameter="Home"/>
-                <Button Grid.Column="1" Content="Options"/>
-            </Grid>
-            
-            <!--Group List-->
-            <ListView x:Name="GroupList" Grid.Row="1" ItemsSource="{Binding GroupList}">
-                <ListView.ItemTemplate>
-                    <DataTemplate>
-                        <TextBlock Text="{Binding Path=Name}"/>
-                    </DataTemplate>
-                </ListView.ItemTemplate>
-
-                <i:Interaction.Triggers>
-                    <i:EventTrigger EventName="MouseLeftButtonUp">
-                        <i:InvokeCommandAction Command="{Binding ListBoxSelectedCommand}" CommandParameter="{Binding ElementName=GroupList,Path=SelectedItem}"/>
-                    </i:EventTrigger>
-                </i:Interaction.Triggers>
-            </ListView>
-        </Grid>
-        <ContentControl Grid.Column="1" Content="{Binding MidViewModel}"/>
-        
-    </Grid>
-</Window>
